feat: Sprint 7 ecosystem build-out, flagship features, and launch-readiness hardening
- Official website (apps/website), demo playground rate-limiting, plugin
  marketplace skeleton with a real subprocess execution sandbox
  (apps/gateway/plugins/sandbox.py), provider certification checker
  (`setu certify`), and Go/Rust/Java/C# client SDKs alongside the existing
  Python and TypeScript ones.
- Terraform provider, Helm chart, and Kubernetes Operator for declarative
  deployment and lifecycle management - each verified against a real
  environment (a live gateway, `helm template`/`helm lint`, and a local
  `kind` cluster respectively), not just compiled.
- Flagship features: Enterprise Policy Engine (provider allow/denylist,
  context-window floor, secret-pattern blocking, enforced before routing),
  AI Cost Optimizer (cheaper-model recommendations from real usage), AI
  Traffic Replay (concurrent batch replay of historical requests against a
  candidate provider), and AI Gateway Studio (routing rule builder, policy
  simulator, cross-provider prompt tester) in the dashboard.
- Community dashboard page (live GitHub activity), governance/maintainers
  docs, a staged public roadmap, and a growth plan.
- CI dependency-vulnerability scanning for the four new language ecosystems
  (govulncheck, cargo audit, dotnet list package --vulnerable, OWASP
  dependency-check) plus a PR-diff-wide dependency review - this caught and
  fixed real, exploitable CVEs in the Terraform provider's and Operator's
  transitive Go dependencies.
- Every Sprint 7 feature wired into the documentation site's navigation
  (12 new pages), with a real gap in .mintignore fixed along the way
  (apps/website and the Helm chart's templates were breaking `mint validate`
  since their respective epics landed).
- Deployment-ready packaging for the website and marketplace registry
  (Dockerfiles, a Caddy reverse-proxy example with automatic HTTPS, a
  production compose overlay) and a DEPLOYMENT_RUNBOOK.md - stops short of
  creating hosting accounts or registering a domain, which stay a
  deliberate, user-driven decision.
- An honest launch-readiness self-audit (LAUNCH_READINESS.md) with gaps
  labeled and updated as they closed, rather than a rubber stamp.
- Fixed two real bugs found during final verification: a flaky SQLite
  shared-cache locking race in two CRUD-lifecycle tests (tests/conftest.py's
  new drain-then-retry helpers), and a ruff SIM105 violation in the plugin
  sandbox that would have failed the gating python-lint CI job.

+++ apps/gateway/models/catalog.py
@@ -28,7 +28,7 @@ class ModelDefinition(BaseModel):
         default=0.0, ge=0.0, description="Approximate USD cost per 1K output tokens (0 for self-hosted/free models)"
     )
 
-    def estimate_cost(self, input_tokens: int, output_tokens: int = 0) -> float:
+    def estimate_cost(self, input_tokens: float, output_tokens: float = 0) -> float:
         """Estimate request cost in USD from token counts. Pricing is illustrative and
         should be kept in sync with each provider's published pricing page."""
         return round(
